Refactor: 대시보드 조회를 hook으로 변경
